Complete all tasks
I answer all of the question at the end of each tasks. Fixed the variable according to the new update on studip. Fixed some axis label. Fixed minor error in function of task 3 and 4. One of them is about including the offset, which has already been included in sklearn, so we dont have to do it again.